add si4432 class,implement it. change main.cpp to test the si4432 class(send message and recv message). Note,unitl now,the code for wait si4432 interrupt is just wait interrupt,it may lead to infinite loop.

+++ main/src/unit.cpp
@@ -33,7 +33,7 @@ void Unit::loadCurrentMeters()
     if (false == waitingLine.empty())
     {
         list<int>::size_type sz = waitingLine.size();
-        for (int i = 0; (i < 4) && (i < sz); i++)
+        for (int i = 0; (i < 4) && (i < (int)sz); i++)
         {
             int tmp;
             tmp = waitingLine.front();
